% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{check_loss}
\alias{check_loss}
\title{Compute Check loss and its gradient}
\usage{
check_loss(tau = 0.5)
}
\arguments{
\item{tau}{A numeric value between 0 and 1 specifying the quantile to estimate. Defaults to 0.5.}
}
\value{
A function performs value (\code{loss}) and its gradients (\code{grad}) of the \code{tau}-th quantile loss function.
}
\description{
This function defines the function of the check loss and its gradient for a given value \code{tau}.
The check loss is a robust loss function that is commonly used as a loss function in quantile regression problem.
}
\examples{
u <- rnorm(10)
loss <- check_loss()
loss_result <- loss(u)
loss_result$loss # loss value
loss_result$grad # gradients
}
