% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emp_kernel_map.R
\name{emp_kernel_map}
\alias{emp_kernel_map}
\title{Empirical kernel mapping matrix}
\usage{
emp_kernel_map(train_x, landmarks, kernel, q = NULL, scale = T)
}
\arguments{
\item{train_x}{Training data (covariates). It should be a numeric matrix.}

\item{landmarks}{A set of indices from \code{train_x} to use as landmarks. It should be a numeric vector.}

\item{kernel}{A kernel generating function provided in the \code{kernlab} package.}

\item{q}{Cumulative proportion of eigenvalues to retain (for dimensionality reduction). Defaults to \code{NULL} (use full rank). It should be a numeric value between 0 and 1.}

\item{scale}{A logical value indicating whether to standardize the data. Defaults to \code{TRUE}.}
}
\value{
A function generates empirical kernel mapping matrix of \eqn{k(}\code{train_x}, \code{test_x}\eqn{)}.
\subsection{Arguments}{
\itemize{
\item \code{test_x} Test data (covariates). It should be a numeric matrix.
}
}

\subsection{Value}{
\itemize{
\item Empirical kernel mapping matrix as a numeric matrix.
}
}
}
\description{
This function generates an empirical kernel mapping function based on the provided data and kernel.
The function uses a subset of training data to approximate the full kernel matrix, improving computational efficiency.
}
\examples{
library(kernlab)
rbfkernel <- rbfdot(sigma = 0.1)

# Fit empirical kernel map
data(iris)
landmarks <- sample(1:100, 10)
emp_kernel <- emp_kernel_map(iris[1:100, 1:4], landmarks, rbfkernel, q=.95)

# Empirical kernel mapping matrices
F_train <- emp_kernel() # empirical kernel mapping for training data
F_test <- emp_kernel(iris[101:150, 1:4]) # empirical kernel mapping for test data
}
