% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{eps_insensitive_loss}
\alias{eps_insensitive_loss}
\title{Compute Epsilon-insensitive loss and its gradient}
\usage{
eps_insensitive_loss(eps = 0.1)
}
\arguments{
\item{eps}{A positive numeric value specifying the threshold for the epsilon-insensitive region. Defaults to 0.1.}
}
\value{
A function performs value (\code{loss}) and its gradients (\code{grad}) of the epsilon insensitive loss function with \code{eps} threshold.
}
\description{
This function defines the function of the epsilon-insensitive loss and its gradient for a given value \code{eps}.
The epsilon insensitive loss is commonly used as a loss function in support vector regression problems.
It ignores errors smaller than a specified epsilon value (\code{eps}).
}
\examples{
u <- rnorm(10)
loss <- eps_insensitive_loss(0.5)
loss_result <- loss(u)
loss_result$loss # loss value
loss_result$grad # gradients
}
