% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{hinge_loss}
\alias{hinge_loss}
\title{Compute Hinge loss and its gradient}
\usage{
hinge_loss(s = 1)
}
\arguments{
\item{s}{A numeric value specifiying the margin parameter. Defaults to 1.}
}
\value{
A function performs value (\code{loss}) and its gradients (\code{grad}) of the hinge loss function.
}
\description{
This function defines the function of the hinge loss and its gradient.
The hinge loss is commonly used as a loss function in support vector machine (SVM) classification problems.
}
\examples{
u <- rnorm(10)
loss <- hinge_loss()
loss_result <- loss(u)
loss_result$loss # loss value
loss_result$grad # gradients
}
