% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{logistic_loss}
\alias{logistic_loss}
\title{Compute Logistic loss and its gradient}
\usage{
logistic_loss()
}
\value{
A function performs value (\code{loss}) and its gradients (\code{grad}) of the logistic loss function.
}
\description{
This function defines the function of the logistic loss and its gradient.
The logistic loss is commonly used as a loss function in logistic regression,
measuring the difference between the predicted probability and the true label.
}
\examples{
u <- rnorm(10)
loss <- logistic_loss()
loss_result <- loss(u)
loss_result$loss # loss value
loss_result$grad # gradients
}
