% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{squared_loss}
\alias{squared_loss}
\title{Compute Squared loss and its gradient}
\usage{
squared_loss()
}
\value{
A function performs value (\code{loss}) and its gradients (\code{grad}) of the squared loss function.
}
\description{
This function defines the function of the squared loss and its gradient.
The squared loss is commonly used as a loss function in regression problems.
}
\examples{
u <- rnorm(10)
loss <- squared_loss()
loss_result <- loss(u)
loss_result$loss # loss value
loss_result$grad # gradients
}
