% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaler.R
\name{standardScaler}
\alias{standardScaler}
\title{Standardized scaler}
\usage{
standardScaler(data, center = T, scale = T)
}
\arguments{
\item{data}{The base data to be used for scaling. It should be a numeric matrix.}

\item{center}{A logical value indicating whether to center the data. Defaults to \code{TRUE}.}

\item{scale}{A logical value indicating whether to scale the data. Defaults to \code{TRUE}.}
}
\value{
A function that performs standardization based on the provided data.
\subsection{Arguments}{
\itemize{
\item \code{x} The new data to be standardized. It should be a numeric matrix. Defaults to \code{data}.
}
}

\subsection{Value}{
\itemize{
\item The standardized data as a numeric matrix.
}
}
}
\description{
This function is a scaler that standardizes the columns of a numeric matrix based on centers and scalings of the provided data.
Standardization involves centering the data (subtracting the mean) and scaling the data (dividing by the standard deviation).
}
\examples{
# Fit the scaler
data(iris)
scaler <- standardScaler(iris[1:100, 1:4])

# Standardized base data
scaler()

# Scale new data
scaler(iris[101:150, 1:4])
}
